Fix yum repo due to CentOS EOL, refactor to use multi-stage builds
Modified the Dockerfile so that the image is built in stages. This allows for smaller container sizes if only certain features are needed, and also allows for a development container for testing new features without needing to prep it every time.

Added a fix to the base container for yum repos. CentOS was EOL on 2022-02-01, so the default online package repos no longer work.

Closes #11 (yum repo list)