Fix argument expansion in wrapper script
The arguments passed to Docker were not being expanded correctly. Fixed this by adding more variables dedicated to function parameters, and use ```eval``` to execute the command rather than running it directly.

This does bring up some security concerns since it would be non-trivial to execute arbitrary code using specially crafted arguments, but it is expected that people using the container are running code on their host machine anyways so this is a conversation for a different day.

See #8 (wrapper versatility)