Fix dnf repo mirror, fix cppcheck installation, add workingdir to PATH
The dnf mirrors that were fixed back in v0.4 apparently are known for poor uptime and performance, and seemed to be the cause of build failures. This was mostly happening in GitHub actions, but started happening locally too. Found a solution online which uses a different mirror.

Fixed the installation of cppcheck. It was not correctly linking to its configuration files, so added them.