Fix gcc compile issue
Previous commits got gcc to compile most of the way, but not quite. Reached out to CAEN devs, they had some awesome suggestions on how to get this working. Tried the following in order:
- Tried building with ```--disable-libsanitizers```, which failed
- Tried building with ```--disable-libsanitizers``` and ```--enable-languages=c,c++,fortran```, which worked!
- Tried building with  ```--enable-languages=c,c++,fortran```, which worked!

So it seems that the existing patch was actually working, but the gcc installer is not doing the best job of figuring out which languages will work.

Tested on some projects which compiled in v0.5 of this container but not on CAEN, and it didn't compile! Very good stuff!

See also: #4