Improve support for non-latest container versions, update README, improve testing CICD
Improve script support for non-latest container versions. This makes it easier to work on testing versions of the container, and stops the script for checking for script updates when developing. This is also needed in order to run the script in GitHub actions, so it doesn't try to pull an existing container rather than using the one built in the runner.

Updated the README docs to reflect some of the more recent changes.

 # Subject line, try to keep under 50 characters