Refactor script to track docker daemon state
The script used to check to make sure the docker daemon was running before doing anything else, and would exit 1 if it wasn't ready. This worked nicely but I want users to be able to update the script and hopefully get updates in the event that docker or the script has a bug.

Changed the startDocker script to just verify it is running and set an env variable. Then other scripts can selectively verify whether Docker is running. This takes more writing, but is a lot more flexible.