Use more portable post-install command in devcontainer
Windows breaks 'initializeCommand' by running the command with 'cmd.exe' locally rather than trying to run it in the container. This is apparently what the command is supposed to do which is not what I realized originally, so changed it to postCreateCommand which specifically runs in the container as intended.